+++ pkg/drivers/common/vmnet/vmnet_stub.go
