+++ pkg/drivers/ssh/ssh.go
@@ -35,7 +35,7 @@ import (
 	"github.com/docker/machine/libmachine/state"
 	"github.com/pkg/errors"
 	"k8s.io/klog/v2"
-	pkgdrivers "k8s.io/minikube/pkg/drivers"
+	"k8s.io/minikube/pkg/drivers/common"
 	"k8s.io/minikube/pkg/minikube/command"
 	"k8s.io/minikube/pkg/minikube/cruntime"
 	"k8s.io/minikube/pkg/minikube/sysinit"
@@ -45,7 +45,7 @@ import (
 // https://minikube.sigs.k8s.io/docs/reference/drivers/ssh/
 type Driver struct {
 	*drivers.BaseDriver
-	*pkgdrivers.CommonDriver
+	*common.CommonDriver
 	EnginePort int
 	SSHKey     string
 	runtime    cruntime.Manager
